+++ CHANGELOG.md
@@ -15,6 +15,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Agent.output_schema` for setting an output schema to be used on the Agent's Prompt Task.
 - `BasePromptDriver.structured_output_strategy` for changing the Structured Output strategy between `native`, `tool`, and `rule`.
 
+### Changed
+- Task bitshift operators can now take a list of Tasks.
+
 ### Fixed
 
 - Occasional crash during `FuturesExecutorMixin` cleanup.